refactor(scanner): Have a separate network module
